requirements.txt
Flask==0.10.1
Ghost.py==0.1b5
Jinja2==2.7.3
MarkupSafe==0.23
PySide==1.2.2
Werkzeug==0.9.6
beautifulsoup4==4.3.2
itsdangerous==0.24
requests==2.5.0
wsgiref==0.1.2
